backend/requirements.txt
asgiref==3.11.0
Django==6.0.1
dnspython==2.6.1
mysql==0.0.3
mysql-connector-python==8.4.0
mysqlclient==2.2.4
pymongo==4.7.2
sqlparse==0.5.5
tzdata==2025.3
djangorestframework==3.15.2
django-cors-headers==4.4.0
pillow==10.4.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9
whitenoise==6.7.0
